+++ rust/Cargo.toml
@@ -4,6 +4,5 @@ members = [
   "agama-dbus-server",
   "agama-derive",
   "agama-lib",
-  "agama-locale-data",
-  "agama-network"
+  "agama-locale-data"
 ]
