+++ rust/agama-dbus-server/src/network/error.rs
