--- rust/agama-network/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "agama-network"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-zbus = "3.12.0"
-agama-lib = { path="../agama-lib" }
-async-std = { version = "1.12.0", features = ["attributes"] }
-thiserror = "1.0.40"
-uuid = { version = "1.3.3", features = ["v4"] }
-parking_lot = "0.12.1"
